Fixed Library Sync author lookup
The replace filter of ". " (period)(space) would fail to catch some authors whose directory name would be otherwise marked incorrect, it also seemed to have an effect on author names that were known good as it would find the author name matched in GoodReads but then would state that the author was not found in the database. This change fixes the issue and allows author import to work correctly.

Please note: I am not picking up the project, I am just a hobbyist that wanted to use this project and wanted to get it working again. I may contribute other patches in the future if I need to create any more.