Fix rename of ini file
Only when the very FIRST save happens, the lazylibrarian.ini file is created. However it tries to rename the current ini file to ini.bak but this doesn't exist so the 'try' errors and doesn't rename the ini.new file back to just ini.

This has knock-on affects for anyone specifying the ini file in their startup scripts.