Fixed OPF file writing (and added ratings to the files)
The sqlite3 row object doesn't implement the right methods to allow the 'in' operator to work.  So all of the optional fields were never written out.

I added a hack to turn the row object into a normal map so that things work properly.

I also added support for writing out the ratings.