Added flags to govern whether the autoadd files are copied
Before the auto-add directory for books and magazines used the same
copying flag as the import.  So if you added new books and wanted to
make a copy for calibre you could either keep all of the raw files
from the downloader and keep the copy in your library, or you could
delete the copy from the downloader and delete the copy from the
library.

I found this behavior annoying, so I added new flags to allow you to
govern the auto-add directory copying separately.

These flags default to 1 so that the files are copied and the file in
lazylibrarian is preserved.