Removed escaping from placeholder parameters in SQL
The code was effectively double escaping ... in some places.  It would
escape " to "" and pass it as a placeholder.  Then the DB code would
further escape that.  So 'E.E. "Doc" Smith' would not work corretly.

I looked at all DB commands across the code and fixed up any
placeholder related problems that I saw.  PRAGMA statements can not
have placeholders so I left them as they were.

There was also a change to a match, we were doing the same replacement
before we tried a match... but the response from the DB should not
have been escaped, so I stripped that out too.

Fix the params